Reformat tables and figures
This change reformats the tables and figures such that they either
render in GitHub markdown or are replaced with TODOs for later cleanup.
This formatting is not as clean as the original, but it does render
enough to be readable. Future cleanups will be needed to add anchors and
titles.

Signed-off-by: Garret Kelly <gdk@google.com>